Changes to have likes, comments, shares count on the post table.

+++ src/main/java/com/annular/filmhook/webmodel/CommentOutputWebModel.java
@@ -19,7 +19,7 @@ public class CommentOutputWebModel {
 	private Integer parentCommentId; // Comment primary key. If comment is for any parent comment.
 
 	private Integer totalLikesCount;
-	private Long totalCommentCount;
+	private Integer totalCommentCount;
 
 	private Boolean status;
 	private String time;
